+++ github/blacklistedRepos.js
@@ -1,0 +1,2 @@
+// kontribusi ke repo ini tidak dianggap
+module.exports = ["programmingCourse", "bellshadebot", "open-api"];
