+++ github/hasNextPage.js
@@ -1,0 +1,7 @@
+const getPageLinks = require("./getPageLinks");
+
+const hasNextPage = (link) => {
+  return getPageLinks(link).next;
+};
+
+module.exports = hasNextPage;
